[MV] Model -> pkg model
