--- pkg/database/model.go
@@ -1,16 +1,0 @@
-package database
-
-type User struct {
-	ID       int64
-	Username string
-	Password string
-}
-
-type Updater struct {
-	ID       int64
-	Username string
-	Password string
-	Domain   string
-	IP       string
-	Interval int64
-}
