--- pkg/ddns/model.go
@@ -1,8 +1,0 @@
-package ddns
-
-type UpdateRequest struct {
-	Username string `json:"username"`
-	Password string `json:"password"`
-	Domain   string `json:"domain"`
-	IP       string `json:"ip"`
-}
